pyproject.toml
[project]
name = "cirno-4"
version = "0.1.0"
description = "cirno-4"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
plugins = [
    "nonebot_plugin_gocqhttp",
    "nonebot_plugin_analysis_bilibili",
    "nonebot_plugin_eventexpiry",
    "nonebot_plugin_githubcard"
]
plugin_dirs = [
    "cirno-4/plugins"
]
builtin_plugins = []

[tool.poetry]
name = "cirno-4"
version = "0.1.0"
description = ""
authors = ["Summerkirakira <summerkirakira@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = {extras = ["fastapi"], version = "^2.0.0"}
nonebot-adapter-onebot = "^2.2.3"
nonebot-plugin-gocqhttp = "^0.6.11"
nonebot-plugin-analysis-bilibili = "^2.5.0"
nonebot-plugin-eventexpiry = "^0.1.1"
nonebot-plugin-githubcard = "^0.1.5"
nonebot-plugin-apscheduler = "^0.3.0"
sqlalchemy = "^2.0.19"
cryptography = "^41.0.2"
nonebot-plugin-localstore = "^0.5.0"
aiopg = "^1.4.0"
aiomysql = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
